+++ Source/BeatBloks.cpp
@@ -482,7 +482,7 @@ void BeatBloks::UpdateZoomExtents()
    mClipEndSlider->SetExtents(mZoomStart, mZoomEnd);
 }
 
-void BeatBloks::OnClicked(int x, int y, bool right)
+void BeatBloks::OnClicked(float x, float y, bool right)
 {
    IDrawableModule::OnClicked(x, y, right);
