+++ Source/CanvasTimeline.h
@@ -57,7 +57,7 @@ public:
    bool MouseScrolled(int x, int y, float scrollX, float scrollY) override;
 
 private:
-   void OnClicked(int x, int y, bool right) override;
+   void OnClicked(float x, float y, bool right) override;
    void GetDimensions(float& width, float& height) override
    {
       width = mWidth;
