+++ Source/NoteTable.cpp
@@ -270,7 +270,7 @@ void NoteTable::DrawModule()
    ofPopStyle();
 }
 
-void NoteTable::OnClicked(int x, int y, bool right)
+void NoteTable::OnClicked(float x, float y, bool right)
 {
    IDrawableModule::OnClicked(x, y, right);
