+++ Source/SamplerGrid.cpp
@@ -355,7 +355,7 @@ void SamplerGrid::GetModuleDimensions(float& width, float& height)
    }
 }
 
-void SamplerGrid::OnClicked(int x, int y, bool right)
+void SamplerGrid::OnClicked(float x, float y, bool right)
 {
    IDrawableModule::OnClicked(x, y, right);
