+++ Source/SeaOfGrain.cpp
@@ -310,7 +310,7 @@ void SeaOfGrain::ButtonClicked(ClickButton* button)
       LoadFile();
 }
 
-void SeaOfGrain::OnClicked(int x, int y, bool right)
+void SeaOfGrain::OnClicked(float x, float y, bool right)
 {
    IDrawableModule::OnClicked(x, y, right);
 }
