+++ Source/TitleBar.h
@@ -145,7 +145,7 @@ private:
    void DrawModuleUnclipped() override;
    bool Enabled() const override { return true; }
    void GetModuleDimensions(float& width, float& height) override;
-   void OnClicked(int x, int y, bool right) override;
+   void OnClicked(float x, float y, bool right) override;
    bool MouseMoved(float x, float y) override;
 
    bool HiddenByZoom() const;
